🛸🚂 ↝ [SSM-139 SSM-172 SSM-171 SSC-31 SSM-170 SSP-67 SSG-185]: Lightning storms coming up
